pyproject.toml
[project]
name = "icd11foundation"
version = "0.1.0"
description = "Preprocessed ICD-11 Foundation source for Mondo ingest"
requires-python = ">=3.10"
dependencies = [
    "linkml>=1.10.0",
    "linkml-owl==0.4.2",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rdflib>=6.0.0",
    "requests>=2.28.0",
    "ruff>=0.8.0",
    "urllib3>=1.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts", "src/icd11foundation"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
